V1.0.1
Complete turn around of the user check for creating channels. It now first checks if the user is banned, if not then it checks if the user is in the server. For both instances; if they are banned or not in the server it WONT log it to prevent spamming